+++ src/example/App.js
@@ -24,9 +24,15 @@ const muiTheme = createMuiTheme({
 });
 
 const rcaTheme = createRcaTheme({
+  info: {
+    background: '#1976d2',
+    color: 'white',
+  },
 });
 
 const rcaSettings = createRcaSettings({
+  // showCloseButton: true,
+  // showActionButton: true,
 });
 
 const AppBase = () => (