+++ src/shared/settings/index.js
@@ -1,5 +1,8 @@
 export default {
   timeout: 5000,
   showCloseButton: false,
+  showActionButton: false,
   enableClickAwayListener: false,
+  actionText: null,
+  actionClickListener: null,
 };