+++ src/shared/theme/index.js
@@ -1,10 +1,22 @@
 import React from 'react';
 import { Info, CheckCircle, Warning, Error } from '@material-ui/icons';
+import ButtonWrapper from './components/ButtonWrapper';
+import ActionButton from './components/ActionButton';
+
+const createButtonWrapper = (child, style) => (
+  <ButtonWrapper child={child} style={style} />
+);
+
+const createActionButton = (text, onClick, style) => (
+  <ActionButton text={text} onClick={onClick} style={style} />
+);
 
 export default {
   info: {
     background: 'white',
     color: '#51525d',
+    button: createActionButton,
+    buttonWrapper: createButtonWrapper,
     adornment: {
       background: '#1976d2',
       color: 'white',
@@ -14,6 +26,8 @@ export default {
   success: {
     background: 'white',
     color: '#51525d',
+    button: createActionButton,
+    buttonWrapper: createButtonWrapper,
     adornment: {
       background: '#388e3c',
       color: 'white',
@@ -23,6 +37,8 @@ export default {
   warning: {
     background: 'white',
     color: '#51525d',
+    button: createActionButton,
+    buttonWrapper: createButtonWrapper,
     adornment: {
       background: '#f57c00',
       color: 'white',
@@ -32,6 +48,8 @@ export default {
   error: {
     background: 'white',
     color: '#51525d',
+    button: createActionButton,
+    buttonWrapper: createButtonWrapper,
     adornment: {
       background: '#d32f2f',
       color: 'white',